// libs-anilist.old/anilist/src/generated/types.ts
export interface searchSeries_Page_media_coverImage {
  __typename: "MediaCoverImage";
  /**
   * The cover image url of the media at a large size
   */
  large: string | null;
}

export interface searchSeries_Page_media_startDate {
  __typename: "FuzzyDate";
  /**
   * Numeric Year (2017)
   */
  year: number | null;
  /**
   * Numeric Month (3)
   */
  month: number | null;
  /**
   * Numeric Day (24)
   */
  day: number | null;
}

export interface searchSeries_Page_media_endDate {
  __typename: "FuzzyDate";
  /**
   * Numeric Year (2017)
   */
  year: number | null;
  /**
   * Numeric Month (3)
   */
  month: number | null;
  /**
   * Numeric Day (24)
   */
  day: number | null;
}

export interface searchSeries_Page_media_title {
  __typename: "MediaTitle";
  /**
   * The romanization of the native language title
   */
  romaji: string | null;
  /**
   * The official english title
   */
  english: string | null;
  /**
   * Official title in it's native language
   */
  native: string | null;
}
// ...
export interface searchSeries_Page_media_relations_edges {
  __typename: "MediaEdge";
  /**
   * The type of relation to the parent model
   */
  relationType: MediaRelation | null;
  node: searchSeries_Page_media_relations_edges_node | null;
}

export interface searchSeries_Page_media_relations {
  __typename: "MediaConnection";
  edges: (searchSeries_Page_media_relations_edges | null)[] | null;
}

export interface searchSeries_Page_media {
  __typename: "Media";
  /**
   * The id of the media
   */
  id: number;
  /**
   * The current releasing status of the media
   */
  status: MediaStatus | null;
  /**
   * The format the media was released in
   */
  format: MediaFormat | null;
  /**
   * The amount of episodes the anime has when complete
   */
  episodes: number | null;
  /**
   * The general length of each anime episode in minutes
   */
  duration: number | null;
  /**
   * Short description of the media's story and characters
   */
  description: string | null;
  /**
   * The cover images of the media
   */
  coverImage: searchSeries_Page_media_coverImage | null;
  /**
   * The first official release date of the media
   */
  startDate: searchSeries_Page_media_startDate | null;
  /**
   * The last official release date of the media
   */
  endDate: searchSeries_Page_media_endDate | null;
  /**
   * The official titles of the media in various languages
   */
  title: searchSeries_Page_media_title | null;
  /**
   * Other media in the same or connecting franchise
   */
  relations: searchSeries_Page_media_relations | null;
}

export interface searchSeries_Page {
  __typename: "Page";
  media: (searchSeries_Page_media | null)[] | null;
}

export interface searchSeries {
  Page: searchSeries_Page | null;
}

export interface searchSeriesVariables {
  query: string;
  type: MediaType;
}

/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: searchUserByName
// ====================================================

export interface searchUserByName_User_avatar {
  __typename: "UserAvatar";
  /**
   * The avatar of user at its largest size
   */
  large: string | null;
  /**
   * The avatar of user at medium size
   */
  medium: string | null;
}

export interface searchUserByName_User {
  __typename: "User";
  /**
   * The id of the user
   */
  id: number;
  /**
   * The name of the user
   */
  name: string;
  /**
   * The user's avatar images
   */
  avatar: searchUserByName_User_avatar | null;
  /**
   * The user's banner images
   */
  bannerImage: string | null;
}

export interface searchUserByName {
  /**
   * User query
   */
  User: searchUserByName_User | null;
}

export interface searchUserByNameVariables {
  name: string;
}
// ...
export enum MediaFormat {
  MANGA = "MANGA",
  MOVIE = "MOVIE",
  MUSIC = "MUSIC",
  NOVEL = "NOVEL",
  ONA = "ONA",
  ONE_SHOT = "ONE_SHOT",
  OVA = "OVA",
  SPECIAL = "SPECIAL",
  TV = "TV",
  TV_SHORT = "TV_SHORT",
}
// ...
export enum MediaStatus {
  CANCELLED = "CANCELLED",
  FINISHED = "FINISHED",
  HIATUS = "HIATUS",
  NOT_YET_RELEASED = "NOT_YET_RELEASED",
  RELEASING = "RELEASING",
}

/**
 * Media type enum, anime or manga.
 */
export enum MediaType {
  ANIME = "ANIME",
  MANGA = "MANGA",
}